DRF/admin_dashboard/serializers.py
from rest_framework import serializers
from core.models import User

class AdminLoginSerializer(serializers.Serializer):
    email = serializers.EmailField()
    password = serializers.CharField(write_only=True)
